refactor: move Dropdown component to layout directory for better structure
